Include Emotion CSS in migration recommendations (#193)
Added recommendation for Emotion CSS as a migration option.

I won't put this above `vanilla-extract`, since that's your choice - but
if someone is already invested in styled-components, it's the most
natural place to go, and the least effort: the API is near-identical to
styled-components, and it's noticeably faster.

We migrated a 400K LOC project in a few days, and it was worth it. 🙂